Read received messages even after the websocket is cancelled
There might be messages remaining in the incoming channel, that have not been read when the websocket coroutine is cancelled. Previously those messages were lost.

This change introduces an independent coroutine that reads messages from the websocket, which enables it to read messages even after the websocket has been closed. The websocket coroutine is enslaved to the receive coroutine, meaning that completion of the receive coroutine will lead to the cancellation of the websocket coroutine.

This solves the problem that invalid socket connections (e.g. improper URL) where not properly handled and eventually timed out.